backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mongodb==4.6.1
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytz==2024.1
email-validator==2.1.0
aiosmtplib==3.0.1
httpx==0.25.2
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0
